+++ src/CMakeLists.txt
@@ -96,7 +96,6 @@ configure_file(config.h.in config.h)
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 add_subdirectory(bgw)
-add_subdirectory(chunk_append)
 add_subdirectory(net)
 add_subdirectory(telemetry)
 add_subdirectory(loader)
