+++ src/nodes/CMakeLists.txt
@@ -5,4 +5,5 @@ set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/chunk_insert_state.c
     ${CMAKE_CURRENT_SOURCE_DIR}/hypertable_insert.c)
 target_sources(${PROJECT_NAME} PRIVATE ${SOURCES})
+add_subdirectory(chunk_append)
 add_subdirectory(constraint_aware_append)
