+++ src/nodes/chunk_append/exec.c
@@ -23,10 +23,10 @@
 
 #include <math.h>
 
-#include "chunk_append/exec.h"
-#include "chunk_append/chunk_append.h"
-#include "chunk_append/explain.h"
-#include "chunk_append/planner.h"
+#include "nodes/chunk_append/exec.h"
+#include "nodes/chunk_append/chunk_append.h"
+#include "nodes/chunk_append/explain.h"
+#include "nodes/chunk_append/planner.h"
 #include "loader/lwlocks.h"
 
 #define INVALID_SUBPLAN_INDEX -1
