+++ tsl/src/nodes/CMakeLists.txt
@@ -1,5 +1,7 @@
-set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/data_node_dispatch.c
-            ${CMAKE_CURRENT_SOURCE_DIR}/data_node_copy.c)
+set(SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/async_append.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/data_node_dispatch.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/data_node_copy.c)
 target_sources(${TSL_LIBRARY_NAME} PRIVATE ${SOURCES})
 add_subdirectory(compress_dml)
 add_subdirectory(decompress_chunk)
